index
information about the car was added specifically when it was produced, the engine , the inside of the car. ALso added where this car is featured in and a small description of the feature. Fixed the slideshow to be continuous instead of pausing a little bit to the background and restarting the slideshow.